# swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"API for Blogging application",
        "title":"Blogging API"
    },
    "servers":[
        {
            "url":"http://localhost:3200",
            "description": "Local server"
        }
    ],
    "components":{
    "securitySchemes":{
        "JWT":{
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    }
    },
    "paths":{
...
        "/api/blogs": {
            "get":{
                "tags": ["Blogs"],
                "summary":"Get Blogs",
                "description":"User will get all blogs",
                "security":[{"JWT": {} }],
                "responses":{
                    "200": {
                        "description":"OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post":{
                "tags": ["Blogs"],
                "summary":"Create Blogs",
                "description":"Create Blogs",
                "security":[{"JWT": {} }],
                "requestBody":{
                    "content":
                    {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title":{
                                        "type": "string"
                                    },
                                    "content":{
                                        "type": "string"
                                    },
                                    "author":{
                                        "type": "string"
                                    }
                                }
                            }
                        } 
                    }
                },
                "responses":{
                    "200": {
                        "description":"OK"
                    },
                    "400": {
                        "description": "Incorrect Data"
                    }
                }
            }
        },
...
    }
}